Fix count bug in SDL_ConvertStereo
The function SDL_ConvertStereo incorrectly computed the number of
samples to duplicate by dividing the size of the source buffer (in
bytes) by two, instead of by the sample size.  Obviously this only
works for 16 bit samples.  Fix the code to divide by the sample size
instead, to make it work for 8-bit and 32-bit samples as well.

Also, the accesses to the source buffer were off by one, since the
pointer was dereferenced before it was decremented instead of after.