Fix rounding error in fractional rate conversions
When doing a rate conversion where the destination rate is not a multiple
of the source rate (this includes e.g. 2x downsample), it is imperative
to make sure that the calculated destination size (in bytes) is a multiple
of the sample size multiplied by the number of channels, and round it
downwards if it is not.  Failure to do so not only results in a bogus
resulting len_cvt, but also causes the rate conversion loop to write outside
the allocated buffer.

After fixing sdlgenaudiocvt.pl, regenerated SDL_audiotypecvt.c.